+++ artifacts/omniproject/src/pages/Settings.tsx
@@ -32,6 +32,7 @@ import { CostRulesAdmin } from "../components/settings/CostRulesAdmin";
 import { CustomReportsAdmin } from "../components/settings/CustomReportsAdmin";
 import { CustomBackendAdmin } from "../components/settings/CustomBackendAdmin";
 import { ContentPagesAdmin } from "../components/settings/ContentPagesAdmin";
+import { FederatedPeersAdmin } from "../components/settings/FederatedPeersAdmin";
 import { PriorityWeightsAdmin } from "../components/settings/PriorityWeightsAdmin";
 import { GovernanceRulesAdmin } from "../components/settings/GovernanceRulesAdmin";
 import { ScopeUpliftAdmin } from "../components/settings/ScopeUpliftAdmin";
@@ -433,6 +434,10 @@ export function Settings() {
 
       <div className="mt-10">
         <PriorityWeightsAdmin />
+      </div>
+
+      <div className="mt-10">
+        <FederatedPeersAdmin />
       </div>
 
       <div className="mt-10">